+++ src/app/Playground/VetleSamples/CollectionViewTests/CollectionViewTests.xaml.cs
@@ -16,14 +16,4 @@ public partial class CollectionViewTests
     {
         Shell.Current.Navigation.PushAsync(new GroupedCollectionView());
     }
-
-    private void RegularCollectionView2(object sender, EventArgs e)
-    {
-        Shell.Current.Navigation.PushAsync(new RegularCollectionView2());
-    }
-
-    private void GroupedCollectionView2(object sender, EventArgs e)
-    {
-        Shell.Current.Navigation.PushAsync(new GroupedCollectionView2());
-    }
 }