+++ src/library/DIPS.Mobile.UI/Components/CarouselView/CarouselView2.cs
@@ -1,6 +1,0 @@
-namespace DIPS.Mobile.UI.Components.CarouselView;
-
-/// <summary>
-/// Uses the new handler on iOS
-/// </summary>
-public class CarouselView2 : Microsoft.Maui.Controls.CarouselView;