+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/CollectionView.Properties.cs
@@ -88,11 +88,6 @@ public partial class CollectionView
         {
             handler.ReloadData(handler);
         }
-
-#if __IOS__
-        if(Handler is CollectionView2Handler handler2)
-            handler2.ReloadData(handler2);
-#endif
     }
 }