+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/CollectionView2.cs
@@ -1,9 +1,0 @@
-namespace DIPS.Mobile.UI.Components.Lists;
-
-/// <summary>
-/// Uses the new handler on iOS
-/// </summary>
-public class CollectionView2 : CollectionView
-{
-    
-}