+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [55.0.5]
+- [CollectionView] Remove CollectionView2 and use new default iOS Handler.
+- [CarouselView] Remove CarouselView2.
+
 ## [55.0.4]
 - [ChipGroup] Fixed issue where multiple chips could remain active when SelectedItems binding changed externally in Single selection mode